basic backend structure

+++ backend/controllers/goalController.js
@@ -3,9 +3,10 @@ const Goal = require("../model/goalModel");
 //@desc Get Goals
 //@route GET /api/goals
 //@access Private
-
+const User = require("../model/userModel");
+const userModel = require("../model/userModel");
 const getGoals = asyncHandler(async (req, res) => {
-	const goals = await Goal.find();
+	const goals = await Goal.find({ user: req.user.id });
 	res.status(200).json(goals);
 });
 
@@ -21,6 +22,7 @@ const setGoal = asyncHandler(async (req, res) => {
 	}
 	const goal = await Goal.create({
 		text: req.body.text,
+		user: req.user.id,
 	});
 	res.status(200).json(goal);
 });
@@ -35,6 +37,18 @@ const updateGoal = asyncHandler(async (req, res) => {
 		res.status(400);
 		throw new Error("goal not found");
 	}
+
+	const user = await User.findById(req.user.id);
+	//check for user
+	if (!user) {
+		res.status(401);
+		throw new Error("User not found");
+	}
+	//logged in user matches goal user
+	if (goal.user.toString() !== user.id) {
+		throw new Error("User not authorized");
+	}
+
 	const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
 		new: true,
 	});
@@ -51,6 +65,18 @@ const deleteGoals = asyncHandler(async (req, res) => {
 		res.status(400);
 		throw new Error("goal not found");
 	}
+
+	const user = await User.findById(req.user.id);
+	//check for user
+	if (!user) {
+		res.status(401);
+		throw new Error("User not found");
+	}
+	//logged in user matches goal user
+	if (goal.user.toString() !== user.id) {
+		throw new Error("User not authorized");
+	}
+
 	const deletedGoal = await Goal.findByIdAndDelete(req.params.id, req.body, {
 		new: true,
 	});
